+++ sdk/core/core-rest-pipeline/src/index.ts
@@ -87,3 +87,4 @@ export {
   AuthorizeRequestOnChallengeOptions,
 } from "./policies/bearerTokenAuthenticationPolicy";
 export { ndJsonPolicy, ndJsonPolicyName } from "./policies/ndJsonPolicy";
+export { throttlingRetryStrategy } from "./retryStrategies/throttlingRetryStrategy";